Fix float coordinates

+++ classes/Host.py
@@ -13,8 +13,8 @@ class Host( MHost ):
         config['ID'] = name
         config['Name'] = name
         config['State'] = False
-        config['x'] = int(x)
-        config['y'] = int(y)
+        config['x'] = float(x)
+        config['y'] = float(y)
         with open('config.json', 'r') as f:
             data = json.load(f)
         try:
